requirements.txt
requests
chromadb
PySide6
python-dotenv
